move the contents of system/interrupts into skeleton/init_teletype
