+++ teletype/main.c
@@ -26,7 +26,6 @@
 #include "events.h"
 #include "init_teletype.h"
 #include "init_common.h"
-#include "interrupts.h"
 #include "i2c.h"
 #include "kbd.h"
 // #include "monome.h"
